requirements.txt
pip
setuptools
discord
psycopg2-binary